pyproject.toml
[tool.poetry]
name = "lite-mcp"
version = "2.0.0"
description = "A simple Model Context Protocol (MCP) server framework"
authors = [
    "weibingcs <weibingcs@2345.com>"
]
readme = "README.md"
packages = [
    {include = "src"},
]

[tool.poetry.dependencies]
python = "^3.12"
fastmcp = "^2.9"
click = "^8.1.7"
rich = "^14.0.0"
fastapi = "^0.115.13"
uvicorn = "^0.34.3"
psutil = "^7.0.0"
httpx = "^0.28.1"
PyYAML = "^6.0"
requests = "^2.32.4"
apkutils2 = "^1.0.0"
pycryptodome = "^3.23.0"
pymysql = "^1.1.1"
chardet = "^5.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[tool.poetry.scripts]
testmcp = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
